requirements.txt
openai==2.6.1
openai-agents==0.4.2
python-dotenv==1.2.1